api/requirements.txt

flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
langchain-mistralai==0.0.3
langchain==0.1.4
